refactor: modularize landing page into components (#26)
